# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Forma API
  description: API documentation for the Forma API
  version: 0.1.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: forma_token
  schemas:
    Error:
      type: object
      required: [code]
      properties:
        code: { type: string }
        message: { type: string }
        fields:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field: { type: string }
              code: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    LoginResponse:
      type: object
      required: [session_id]
      properties:
        session_id: { type: string }
    SignupRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        display_name: { type: string, minLength: 1 }
    SignupResponse:
      type: object
      required: [id]
      properties:
        id: { type: string, format: uuid }
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }
    ResendVerificationRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    PasswordResetConfirmRequest:
      type: object
      required: [token, new_password]
      properties:
        token: { type: string }
        new_password: { type: string, minLength: 8 }
    UserProfile:
      type: object
      required: [id, email, display_name]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        verified_at: { type: string, format: date-time, nullable: true }
    UpdateUserProfileRequest:
      type: object
      required: [display_name]
      properties:
        display_name: { type: string }
    WhoamiResponse:
      type: object
      required: [user_id]
      properties:
        user_id: { type: string, format: uuid }
    RegisterFormRequest:
      type: object
      required: [url]
      properties:
        url: { type: string, description: "Google Forms のURLまたはフォームID" }
    RegisterFormResponse:
      type: object
      required: [id]
      properties:
        id: { type: string, format: uuid }
    Form:
      type: object
      required: [id, form_id, title, created_at]
      properties:
        id: { type: string, format: uuid }
        form_id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
    ListFormsResponse:
      type: object
      required: [forms]
      properties:
        forms:
          type: array
          items: { $ref: "#/components/schemas/FormSummary" }
    UpdateFormRequest:
      type: object
      properties:
        title_question_id:
          type: string
          nullable: true
          description: "未指定なら変更なし。null のみ解除。"
...
    ListMembersResponse:
      type: object
      required: [members]
      properties:
        members:
          type: array
          items: { $ref: "#/components/schemas/Member" }
    AddMemberRequest:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [admin, editor] }
    ChangeMemberRoleRequest:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [admin, editor] }
    CreateInviteRequest:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [admin, editor] }
    CreateInviteResponse:
      type: object
      required: [invite_id, expires_at]
      properties:
        invite_id: { type: string, format: uuid }
        expires_at: { type: string, format: date-time }
...
    ListFormInvitesResponse:
      type: object
      required: [invites]
      properties:
        invites:
          type: array
          items: { $ref: "#/components/schemas/FormInvite" }
...
    ListFormQuestionsResponse:
      type: object
      required: [questions]
      properties:
        questions:
          type: array
          items: { $ref: "#/components/schemas/FormQuestion" }
...
    ListFormStatusesResponse:
      type: object
      required: [statuses]
      properties:
        statuses:
          type: array
          items: { $ref: "#/components/schemas/FormStatus" }
    CreateFormStatusRequest:
      type: object
      required: [name, display_order]
      properties:
        name: { type: string }
        color: { type: string, nullable: true }
        display_order: { type: integer }
        is_default: { type: boolean }
...
    SyncResponse:
      type: object
      required: [synced, new_tickets]
      properties:
        synced: { type: boolean }
        newTickets: { type: integer }
        last: { type: string, format: date-time }
security:
  - cookieAuth: []
paths:
...
  /v1/auth/signup:
    post:
      summary: サインアップ
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignupRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignupResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Email already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: Internal server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/login:
    post:
      summary: ログイン
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Email not verified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: Internal server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/logout:
    post:
      summary: ログアウト
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/verify-email:
    post:
      summary: メール認証
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyEmailRequest" }
      responses:
        "204":
          description: No Content
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Token not found or expired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/verify-email/resend:
    post:
      summary: 認証メール再送
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResendVerificationRequest" }
      responses:
        "202":
          description: Accepted
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/password-reset:
    post:
      summary: パスワードリセット
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetRequest" }
      responses:
        "202":
          description: Accepted
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/password-reset/confirm:
    post:
      summary: パスワードリセット確認
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetConfirmRequest" }
      responses:
        "204":
          description: No Content
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Token not found or expired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/me:
    get:
      summary: プロフィール取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    patch:
      summary: プロフィール更新
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserProfileRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      summary: アカウント削除
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/me/password:
    patch:
      summary: パスワード変更
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 8 }
      responses:
        "204":
          description: No Content
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Current password incorrect
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/whoami:
    get:
      summary: 現在のユーザーID取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WhoamiResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/forms:
    get:
      summary: フォーム一覧
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListFormsResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      summary: フォーム登録
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterFormRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegisterFormResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Forms not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/forms/{form_id}:
    parameters:
      - in: path
        name: form_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: フォーム取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Form" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    patch:
      summary: フォーム更新
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateFormRequest" }
      responses:
        "204":
          description: Updated
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/forms/{form_id}/sync:
    post:
      summary: 回答の同期
      parameters:
        - in: path
          name: form_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 同期完了
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/forms/{form_id}/members:
    parameters:
      - in: path
        name: form_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: メンバー一覧取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListMembersResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      summary: メンバー追加
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddMemberRequest" }
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/forms/{form_id}/members/{user_id}:
    parameters:
      - in: path
        name: form_id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: user_id
        required: true
        schema: { type: string, format: uuid }
    put:
      summary: メンバーロール変更
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangeMemberRoleRequest" }
      responses:
        "204":
          description: Updated
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      summary: メンバー削除
      responses:
        "204":
          description: No Content
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/forms/{form_id}/invites:
    parameters:
      - in: path
        name: form_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 招待一覧取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListFormInvitesResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      summary: 招待作成
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInviteRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateInviteResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/forms/{form_id}/invites/{invite_id}:
    parameters:
      - in: path
        name: form_id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: invite_id
        required: true
        schema: { type: string, format: uuid }
    delete:
      summary: 招待削除
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/invites/{invite_id}/accept:
    parameters:
      - in: path
        name: invite_id
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: 招待受諾
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/forms/{form_id}/questions:
    parameters:
      - in: path
        name: form_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 質問一覧取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListFormQuestionsResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/forms/{form_id}/statuses:
    parameters:
      - in: path
        name: form_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: ステータス一覧取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListFormStatusesResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      summary: ステータス作成
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateFormStatusRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FormStatus" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/forms/{form_id}/statuses/{status_id}:
    parameters:
      - in: path
        name: form_id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: status_id
        required: true
        schema: { type: string, format: uuid }
    patch:
      summary: ステータス更新
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateFormStatusRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FormStatus" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      summary: ステータス削除
      responses:
        "204":
          description: No Content
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/tickets:
    get:
      summary: チケット一覧取得
      parameters:
        - in: query
          name: form_id
          required: true
          schema: { type: string, format: uuid }
          description: "フィルタリング用のフォームID"
        - in: query
          name: status_id
          schema: { type: string, format: uuid }
          description: "フィルタリング用のステータスID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListTicketsResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/tickets/{ticket_id}:
    parameters:
      - in: path
        name: ticket_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: チケット詳細取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketDetail" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    patch:
      summary: チケット更新
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTicketRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketDetail" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/tickets/{ticket_id}/histories:
    parameters:
      - in: path
        name: ticket_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: チケット履歴一覧取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                { $ref: "#/components/schemas/ListTicketHistoriesResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }